requirements.txt
speedtest-cli
customtkinter